+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2021.1.1
+
+- Fix: Toggled Clips didn't hide on trigger again
+
 ## 2021.1.0
 
 - First Release